Initial publication of ivoa-entities
this is really just a local collection of JAXB annotated classes
derived from schema that are associated with IVOA standards.

It is derived from a similar library that used to
exist in AstroGrid software. That library had hand annotated classes,
which had JPA annotations as as well - As it is quite a while since the
classes were last used I decided to go back to auto-generation for this
new version to pick up any changes.

Some of the helper classes associated with the previous astrogrid
release have been left in place for when this library might have JPA
annotations added back.